AI-Crop-Disease-Detection/requirements.txt
# ======================================================
# AI Crop Disease Detection — Python Requirements
# ======================================================

# ---- Deep Learning ----
tensorflow==2.13.0
keras==2.13.1

# ---- Data Science ----
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
Pillow==10.0.0
opencv-python==4.8.0.76

# ---- FastAPI Backend ----
fastapi==0.103.1
uvicorn[standard]==0.23.2
python-multipart==0.0.6
aiofiles==23.2.1

# ---- Data Validation ----
pydantic==2.3.0

# ---- Database ----
sqlalchemy==2.0.20
aiosqlite==0.19.0

# ---- Image Processing for GradCAM ----
scipy==1.11.2

# ---- Dataset Download ----
kaggle==1.5.16
requests==2.31.0
tqdm==4.66.1

# ---- Utilities ----
python-dotenv==1.0.0
python-jose==3.3.0
passlib==1.7.4

# ---- Testing ----
pytest==7.4.2
httpx==0.25.0
